Sprint 5 M2: EQ control plane — slider moves are now audible
Wire the 31-band EQ through to the live AU so moving a slider / picking a
preset changes the sound. Control flows: EQViewModel.dispatchAllBands ->
EQSafetyClamp -> AudioViewModel.publishEQGains -> engine -> C-ABI
publishEQBandGains -> computeBiquadCascade (off-RT) -> publishTargetState.

C-ABI bridge (AUAudioUnit.mm + AudioUnitRegistrationBridge.h), per
modern-cplus-plus-expert review:
- publishEQBandGains(auUnit, gainsDb, count, sampleRate): validates (31 bands,
  SR>0, non-null), computes the minimum-phase biquad cascade off-RT, and
  publishes. Borrowed (__bridge) AU pointer, matching publishTargetState.
- The AU retains an authoritative TargetState (_currentState) and EQ updates
  mutate only .eq before re-publishing, so clarity/loudness/limiter keep their
  last-set values instead of reverting to defaults each slider move.

Swift wiring, per swift-expert review:
- publishEQGains on the protocol/engine/viewmodel; design SR read from the AU
  output bus (48k graph fallback). Replaced the dead per-band setParameter loop
  (setAUParameter is a stub) with one publish; removed unused base param ID.
- dispatchAllBands guards on isEngineReady (closes the narrow shutdown race).

EQModuleCoefficients.h: this header is now compiled in the strict production
target for the first time (was test-bridge-only) — fixed 14 inherited
sign-conversion / int->uint8_t warnings with behavior-preserving static_casts.
Verified unchanged via the standalone coefficient test (10/10).

Validation (audio-dsp review = "measure then decide" on coefficient zipper):
- New harness tests: EQ_FrequencyResponseAccuracy (+6 dB @ 1 kHz measured
  within ±1 dB) and EQ_CoefficientSwapNoClick. The swap test MEASURED the
  worst-case flat->+12 dB jump: boundary step 0.024 << settled per-sample
  delta 0.130 => NO audible click. So the vDSP_biquad setup-swap is kept as-is
  (no vDSP_biquadm migration / gain-smoothing needed). Harness 19/19.
- C++ harness 19/19; coefficient test 10/10; app builds.

+++ Sources/AudioDSP/EQ/EQModuleCoefficients.h
@@ -56,7 +56,7 @@ namespace AdaptiveSound
             bool allZero = true;
             for (int i = 0; i < kNumBands; ++i)
             {
-                if (std::abs(gains[i]) > kFlatGainThresholdDb)
+                if (std::abs(gains[static_cast<size_t>(i)]) > kFlatGainThresholdDb)
                 {
                     allZero = false;
                     break;
@@ -86,10 +86,10 @@ namespace AdaptiveSound
             }
 
             // Copy to result
-            result.numBiquads = std::min(numBiquads, kMaxBiquads);
+            result.numBiquads = static_cast<uint8_t>(std::min(numBiquads, kMaxBiquads));
             for (int i = 0; i < result.numBiquads; ++i)
             {
-                result.biquads[i] = biquads[i];
+                result.biquads[static_cast<size_t>(i)] = biquads[static_cast<size_t>(i)];
             }
             result.masterGainLinear = 1.0F;
 
@@ -133,9 +133,9 @@ namespace AdaptiveSound
 
             for (int i = 0; i < kNumBands; ++i)
             {
-                if (std::abs(gains[i]) > kActiveRegionThresholdDb)
+                if (std::abs(gains[static_cast<size_t>(i)]) > kActiveRegionThresholdDb)
                 {
-                    activeRegions[i] = true;
+                    activeRegions[static_cast<size_t>(i)] = true;
                     numActiveRegions++;
                 }
             }
@@ -153,25 +153,25 @@ namespace AdaptiveSound
 
             while (i < kNumBands && numBiquads < kMaxBiquads)
             {
-                if (activeRegions[i])
+                if (activeRegions[static_cast<size_t>(i)])
                 {
                     // Find the peak of this gain region
                     int peakIdx = i;
-                    float peakGain = gains[i];
+                    float peakGain = gains[static_cast<size_t>(i)];
 
                     // Extend region forward
-                    while (i < kNumBands && activeRegions[i])
+                    while (i < kNumBands && activeRegions[static_cast<size_t>(i)])
                     {
-                        if (gains[i] > peakGain)
+                        if (gains[static_cast<size_t>(i)] > peakGain)
                         {
-                            peakGain = gains[i];
+                            peakGain = gains[static_cast<size_t>(i)];
                             peakIdx = i;
                         }
                         i++;
                     }
 
                     // Create peaking biquad at peak frequency with peak gain
-                    float centerFreq = kCenterFrequencies[peakIdx];
+                    float centerFreq = kCenterFrequencies[static_cast<size_t>(peakIdx)];
                     float gainDb = peakGain;
 
                     // Clamp to safe range (±12 dB as per spec)
@@ -180,7 +180,7 @@ namespace AdaptiveSound
                     // Create peaking filter
                     EQParams::BiquadCoeffs biquad =
                         designPeakingFilter(centerFreq, gainDb, sampleRate);
-                    outBiquads[numBiquads++] = biquad;
+                    outBiquads[static_cast<size_t>(numBiquads++)] = biquad;
                 }
                 else
                 {
@@ -255,7 +255,7 @@ namespace AdaptiveSound
         {
             for (int i = 0; i < numBiquads; ++i)
             {
-                const auto& b = biquads[i];
+                const auto& b = biquads[static_cast<size_t>(i)];
 
                 // Check stability: poles must be inside unit circle
                 // For biquad: A(z) = 1 + a1*z^-1 + a2*z^-2
